fixed the next page button issue
if options.totalItems is equal to 0, the length of options.data is greater than 1,  the value of options.data.length < 1 is false, so now the next page button is enabled. My project has come with this isssue, so i have changed  return options.data.length < 1 to return true. now it works!